basic upload interface + pages
In this updated I included all the pages and the routes between them. The interfaces includes buttons to do the different navigations.

On a second note, I built the file upload in the demo page. This page now also includes a way to remove files after loading them and clearning your uploads.