fixed authenticaiton with own backend, removed auth functions and most passport remainders
This commit adds mostyl the connection with the backend as a replacement for the old server that was included for authentication using passport. I now built my own mongoDB authentication and removed almost all remains of next auth, some should still be removed. This can now be merged into master.